Split huge spec file into separate ones for clarity
