+++ spec/types/enum.spec.ts
@@ -1,0 +1,18 @@
+import { z } from 'zod';
+import { expectSchema } from '../lib/helpers';
+
+describe('enum', () => {
+  it('supports enums', () => {
+    const schema = z
+      .enum(['option1', 'option2'])
+      .openapi({ refId: 'Enum', description: 'All possible options' });
+
+    expectSchema([schema], {
+      Enum: {
+        type: 'string',
+        description: 'All possible options',
+        enum: ['option1', 'option2'],
+      },
+    });
+  });
+});
